refactor: move cpfValidator to validation layer
