+++ src/main/factories/controllers/create-person-controller-factory.ts
@@ -1,5 +1,5 @@
 import { CreatePersonController } from '../../../presentation/controllers/create-person/create-person-controller'
-import { CpfValidatorAdapter } from '../../adapters/validators/cpf-validator-adapter'
+import { CpfValidatorAdapter } from '../../../validation/validators/cpf-validator-adapter'
 import { makeCreatePerson } from '../usecases/create-person-factory'
 import { makeLoadPerson } from '../usecases/load-person-factory'
