+++ src/main/factories/controllers/create-relationship-controller-factory.ts
@@ -1,5 +1,5 @@
 import { CreateRelationshipController } from '../../../presentation/controllers/create-relationship/create-relationship-controller'
-import { CpfValidatorAdapter } from '../../adapters/validators/cpf-validator-adapter'
+import { CpfValidatorAdapter } from '../../../validation/validators/cpf-validator-adapter'
 import { makeCreateRelationship } from '../usecases/create-relationship-factory'
 import { makeLoadPerson } from '../usecases/load-person-factory'
