+++ src/main/factories/controllers/load-person-controller-factory.ts
@@ -1,5 +1,5 @@
 import { LoadPersonController } from '../../../presentation/controllers/load-person/load-person-controller'
-import { CpfValidatorAdapter } from '../../adapters/validators/cpf-validator-adapter'
+import { CpfValidatorAdapter } from '../../../validation/validators/cpf-validator-adapter'
 import { makeLoadPerson } from '../usecases/load-person-factory'
 
 export const makeLoadPersonController = (): LoadPersonController => {
