+++ src/validation/protocols/index.ts
@@ -1,0 +1,1 @@
+export * from './cpf-validator'
